backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-jose
passlib[bcrypt]
python-multipart
pydantic-settings